requirements.txt
Flask==2.2.2
Pillow==9.5.0
Pillow==10.3.0
python-dotenv==1.0.1
Requests==2.32.3
spotipy==2.23.0
spotipy==2.24.0
wikipedia==1.4.0
